organize the structure for implementation
