+++ src/animenews/lib/exception.py
