+++ src/animenews/lib/fetch.py
