+++ src/animenews/lib/node.py
